// src/app/models/Advertisement.model.ts
export class AdvertisementModel {
  /*
     "id": 2,
      "userId": 4,
      "dateTime": "21/04/2021 19:20",
      "title": "You like it",
      "description": "Only for this month",
      "discount": 30,
      "urlToImage": "https://www.lasamarillasdezipaquira.com/oc-content/uploads/1/352.jpg",
      "promotion": true
   */
  id: number;
  userId: number;
  dateTime: string;
  title: string;
  description: string;
  discount: number;
  urlToImage: string;
  promotion: boolean;
}